requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
opencv-python-headless
python-multipart
